Add onHold confirmation
