+++ packages/invoices/src/InboundShipment/DetailView/Footer/OnHoldButton.tsx
@@ -1,0 +1,30 @@
+import React, { memo } from 'react';
+import {
+  ToggleButton,
+  useTranslation,
+  useConfirmationModal,
+} from '@openmsupply-client/common';
+import { useInboundFields, useIsInboundDisabled } from '../../api';
+
+export const OnHoldButtonComponent = memo(() => {
+  const t = useTranslation('replenishment');
+  const { onHold, update } = useInboundFields('onHold');
+  const isDisabled = useIsInboundDisabled();
+  const getConfirmation = useConfirmationModal({
+    message: t('messages.on-hold-confirmation'),
+    title: t('heading.are-you-sure'),
+    onConfirm: () => update({ onHold: !onHold }),
+  });
+
+  return (
+    <ToggleButton
+      disabled={isDisabled}
+      value={onHold}
+      selected={onHold}
+      onClick={() => getConfirmation()}
+      label={t('label.hold')}
+    />
+  );
+});
+
+export const OnHoldButton = memo(OnHoldButtonComponent);
